refactor(blocking): move blocking API to `python.blocking` and unify type names
