+++ Cargo.toml
@@ -41,17 +41,17 @@ arc-swap = "1.7.1"
 url = "2.5.4"
 bytes = "1.10.1"
 futures-util = { version = "0.3.31", default-features = false }
-wreq = { version = "6.0.0-rc.7", features = [
+wreq = { version = "6.0.0-rc.8", features = [
     "json",
     "socks",
     "stream",
     "multipart",
+    "cookies",
     "gzip",
     "brotli",
     "deflate",
     "zstd",
     "ws",
-    "cookies",
 ] }
 wreq-util = { version = "3.0.0-rc.1", features = ["emulation-rand"] }
