+++ python/examples/blocking/proxy.py
@@ -1,8 +1,8 @@
-from rnet import BlockingClient
+from rnet.blocking import Client
 
 
 def main():
-    client = BlockingClient()
+    client = Client()
     resp = client.post(
         "https://httpbin.org/anything",
         proxy="http://127.0.0.1:6152",
