+++ python/examples/get.py
@@ -1,11 +1,13 @@
 import asyncio
 import rnet
+from rnet.emulation import Emulation
 
 
 async def main():
     resp = await rnet.get(
         "https://tls.peet.ws/api/all",
         timeout=10,
+        emulation=Emulation.Firefox139,
     )
     print(await resp.text())
