--- src/client/async_impl/response/mod.rs
@@ -1,7 +1,0 @@
-mod http;
-mod ws;
-
-pub use self::{
-    http::{Response, Streamer},
-    ws::{Message, WebSocket},
-};
