--- src/client/blocking/response/mod.rs
@@ -1,7 +1,0 @@
-mod http;
-mod ws;
-
-pub use self::{
-    http::{BlockingResponse, BlockingStreamer},
-    ws::BlockingWebSocket,
-};
