--- src/client/json.rs
@@ -1,15 +1,0 @@
-use indexmap::IndexMap;
-use pyo3::prelude::*;
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, FromPyObject, IntoPyObject, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum Json {
-    Object(IndexMap<String, Json>),
-    Boolean(bool),
-    Number(isize),
-    Float(f64),
-    String(String),
-    Null(Option<isize>),
-    Array(Vec<Json>),
-}
