--- src/client/net.rs
@@ -1,27 +1,0 @@
-use std::fmt;
-
-use pyo3::{IntoPyObjectExt, prelude::*};
-
-/// A IP socket address.
-#[derive(Clone, Copy, PartialEq, Eq)]
-#[pyclass(eq, str)]
-pub struct SocketAddr(pub std::net::SocketAddr);
-
-#[pymethods]
-impl SocketAddr {
-    /// Returns the IP address of the socket address.
-    fn ip<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
-        self.0.ip().into_bound_py_any(py)
-    }
-
-    /// Returns the port number of the socket address.
-    fn port(&self) -> u16 {
-        self.0.port()
-    }
-}
-
-impl fmt::Display for SocketAddr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
