--- src/client/param/mod.rs
@@ -1,5 +1,0 @@
-mod client;
-mod request;
-mod ws;
-
-pub use self::{client::ClientParams, request::RequestParams, ws::WebSocketParams};
