+++ src/client/response/http.rs
@@ -1,16 +1,15 @@
-use std::{ops::Deref, pin::Pin, sync::Arc};
+use std::sync::Arc;
 
 use arc_swap::ArcSwapOption;
-use futures_util::{Stream, TryStreamExt};
 use mime::Mime;
 use pyo3::{IntoPyObjectExt, prelude::*, pybacked::PyBackedStr};
 use pyo3_async_runtimes::tokio::future_into_py;
-use tokio::sync::Mutex;
 use wreq::{Url, header, tls::TlsInfo};
 
+use super::Streamer;
 use crate::{
     buffer::{Buffer, BytesBuffer, PyBufferProtocol},
-    client::{SocketAddr, json::Json},
+    client::{SocketAddr, body::Json},
     error::Error,
     http::{Version, cookie::Cookie, header::HeaderMap, status::StatusCode},
 };
@@ -27,8 +26,14 @@ pub struct Response {
     response: ArcSwapOption<wreq::Response>,
 }
 
+/// A blocking response from a request.
+#[pyclass(name = "Response", subclass)]
+pub struct BlockingResponse(Response);
+
+// ===== impl Response =====
+
 impl Response {
-    /// Create a new `Response` instance.
+    /// Create a new [`Response`] instance.
     pub fn new(mut response: wreq::Response) -> Self {
         Response {
             url: response.url().clone(),
@@ -41,8 +46,7 @@ impl Response {
         }
     }
 
-    /// Consumes the `Response` and returns the inner `wreq::Response`.
-    pub fn inner(&self) -> PyResult<wreq::Response> {
+    fn inner(&self) -> PyResult<wreq::Response> {
         self.response
             .swap(None)
             .and_then(Arc::into_inner)
@@ -191,20 +195,19 @@ impl Response {
     /// Closes the response connection.
     pub fn close<'py>(&'py self, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
         let res = self.inner();
-        future_into_py(py, async move {
-            let _ = res.map(drop);
-            Ok(())
-        })
+        future_into_py(py, async move { res.map(drop) })
     }
 }
 
 #[pymethods]
 impl Response {
+    #[inline]
     fn __aenter__<'py>(slf: PyRef<'py, Self>, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
         let slf = slf.into_py_any(py)?;
         future_into_py(py, async move { Ok(slf) })
     }
 
+    #[inline]
     fn __aexit__<'py>(
         &'py self,
         py: Python<'py>,
@@ -216,81 +219,145 @@ impl Response {
     }
 }
 
-type InnerStreamer = Pin<Box<dyn Stream<Item = wreq::Result<bytes::Bytes>> + Send + 'static>>;
+// ===== impl BlockingResponse =====
 
-/// A byte stream response.
-/// An asynchronous iterator yielding data chunks from the response stream.
-/// Used to stream response content.
-/// Implemented in the `stream` method of the `Response` class.
-/// Can be used in an asynchronous for loop in Python.
-#[derive(Clone)]
-#[pyclass(subclass)]
-pub struct Streamer(Arc<Mutex<Option<InnerStreamer>>>);
-
-impl Deref for Streamer {
-    type Target = Arc<Mutex<Option<InnerStreamer>>>;
-
-    #[inline(always)]
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl Streamer {
-    /// Create a new `Streamer` instance.
-    pub fn new(
-        stream: impl Stream<Item = wreq::Result<bytes::Bytes>> + Send + 'static,
-    ) -> Streamer {
-        Streamer(Arc::new(Mutex::new(Some(Box::pin(stream)))))
+#[pymethods]
+impl BlockingResponse {
+    /// Returns the URL of the response.
+    #[getter]
+    pub fn url(&self) -> &str {
+        self.0.url()
     }
 
-    pub async fn _anext(
-        streamer: Arc<Mutex<Option<InnerStreamer>>>,
-        error: fn() -> PyErr,
-    ) -> PyResult<Py<PyAny>> {
-        let mut lock = streamer.lock().await;
-        let val = lock.as_mut().ok_or_else(error)?.try_next().await;
+    /// Returns the status code of the response.
+    #[getter]
+    pub fn status(&self) -> StatusCode {
+        self.0.status()
+    }
 
-        drop(lock);
+    /// Returns the HTTP version of the response.
+    #[getter]
+    pub fn version(&self) -> Version {
+        self.0.version()
+    }
 
-        let buffer = val
-            .map_err(Error::Request)?
-            .map(BytesBuffer::new)
-            .ok_or_else(error)?;
+    /// Returns the headers of the response.
+    #[getter]
+    pub fn headers(&self) -> HeaderMap {
+        self.0.headers()
+    }
 
-        Python::with_gil(|py| buffer.into_bytes(py))
+    /// Returns the cookies of the response.
+    #[getter]
+    pub fn cookies(&self, py: Python) -> Vec<Cookie> {
+        self.0.cookies(py)
+    }
+
+    /// Returns the content length of the response.
+    #[getter]
+    pub fn content_length(&self) -> u64 {
+        self.0.content_length()
+    }
+
+    /// Returns the remote address of the response.
+    #[getter]
+    pub fn remote_addr(&self) -> Option<SocketAddr> {
+        self.0.remote_addr()
+    }
+
+    /// Encoding to decode with when accessing text.
+    #[getter]
+    pub fn encoding(&self, py: Python) -> String {
+        self.0.encoding(py)
+    }
+
+    /// Returns the TLS peer certificate of the response.
+    pub fn peer_certificate<'py>(
+        &'py self,
+        py: Python<'py>,
+    ) -> PyResult<Option<Bound<'py, PyAny>>> {
+        self.0.peer_certificate(py)
+    }
+
+    /// Returns the text content of the response.
+    pub fn text(&self, py: Python) -> PyResult<String> {
+        py.allow_threads(|| {
+            let resp = self.0.inner()?;
+            pyo3_async_runtimes::tokio::get_runtime()
+                .block_on(resp.text())
+                .map_err(Error::Request)
+                .map_err(Into::into)
+        })
+    }
+
+    /// Returns the text content of the response with a specific charset.
+    pub fn text_with_charset(&self, py: Python, encoding: PyBackedStr) -> PyResult<String> {
+        py.allow_threads(|| {
+            let resp = self.0.inner()?;
+            pyo3_async_runtimes::tokio::get_runtime()
+                .block_on(resp.text_with_charset(&encoding))
+                .map_err(Error::Request)
+                .map_err(Into::into)
+        })
+    }
+
+    /// Returns the JSON content of the response.
+    pub fn json(&self, py: Python) -> PyResult<Json> {
+        py.allow_threads(|| {
+            let resp = self.0.inner()?;
+            pyo3_async_runtimes::tokio::get_runtime()
+                .block_on(resp.json::<Json>())
+                .map_err(Error::Request)
+                .map_err(Into::into)
+        })
+    }
+
+    /// Returns the bytes content of the response.
+    pub fn bytes(&self, py: Python) -> PyResult<Py<PyAny>> {
+        py.allow_threads(|| {
+            let resp = self.0.inner()?;
+            let buffer = pyo3_async_runtimes::tokio::get_runtime()
+                .block_on(resp.bytes())
+                .map(BytesBuffer::new)
+                .map_err(Error::Request)?;
+
+            Python::with_gil(|py| buffer.into_bytes(py))
+        })
+    }
+
+    /// Convert the response into a `Stream` of `Bytes` from the body.
+    pub fn stream(&self, py: Python) -> PyResult<Streamer> {
+        self.0.stream(py)
+    }
+
+    /// Closes the response connection.
+    pub fn close(&self, py: Python) -> PyResult<()> {
+        py.allow_threads(|| self.0.inner().map(drop))
     }
 }
 
 #[pymethods]
-impl Streamer {
-    fn __aiter__(slf: PyRef<Self>) -> PyRef<Self> {
+impl BlockingResponse {
+    #[inline]
+    fn __enter__(slf: PyRef<Self>) -> PyRef<Self> {
         slf
     }
 
-    fn __anext__<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
-        future_into_py(
-            py,
-            Streamer::_anext(self.0.clone(), || Error::StopAsyncIteration.into()),
-        )
-    }
-
-    fn __aenter__<'py>(slf: PyRef<'py, Self>, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
-        let slf = slf.into_py_any(py)?;
-        future_into_py(py, async move { Ok(slf) })
-    }
-
-    fn __aexit__<'py>(
+    #[inline]
+    fn __exit__<'py>(
         &self,
         py: Python<'py>,
         _exc_type: &Bound<'py, PyAny>,
         _exc_value: &Bound<'py, PyAny>,
         _traceback: &Bound<'py, PyAny>,
-    ) -> PyResult<Bound<'py, PyAny>> {
-        let streamer = self.0.clone();
-        future_into_py(py, async move {
-            drop(streamer.lock().await.take());
-            Ok(())
-        })
+    ) -> PyResult<()> {
+        self.close(py)
+    }
+}
+
+impl From<Response> for BlockingResponse {
+    #[inline]
+    fn from(response: Response) -> Self {
+        Self(response)
     }
 }
